linux64: compile filesystem

+++ filesystem/filesystem_stdio/ThreadSafeRefCountedObject.h
@@ -76,7 +76,7 @@ public:
 	}
 
 private:
-	CInterlockedIntT<long> m_RefCount;
+	CInterlockedIntT<int> m_RefCount;
 	T m_pObject;
 };
